test: monkey p. data path fetching fxn instead of data manager fixture
Originally, I had a global test fixture of a mock data manager, but
this interfered with SpecletModel-specific settings of the manager.
The mock version was made so that the tests would use the test data.
The new test pattern is to monkey patch the "data path"-fetching
function to return the path to the test file if any of the files
it is standing in for are requested. This pattern seems to work
much better.